requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
argparse
pathlib
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0 